Folders, files, classes, namespaces, link, routing ... have been renamed to 'registration'.
