+++ src/App/Frontend/Templates/layout.php
@@ -37,7 +37,7 @@
 						<a class="nav-link py-3 px-0 px-lg-3 rounded js-scroll-trigger" href="/news-page-1">Les News</a>
 					</li>
 					<li class="nav-item mx-0 mx-lg-1">
-						<a class="nav-link py-3 px-0 px-lg-3 rounded js-scroll-trigger" href="/inscription">Inscription</a>
+						<a class="nav-link py-3 px-0 px-lg-3 rounded js-scroll-trigger" href="/registration">Inscription</a>
 					</li>
 					<li class="nav-item mx-0 mx-lg-1">
 						<a class="nav-link py-3 px-0 px-lg-3 rounded js-scroll-trigger" href="/#contact">Contact</a>
